requirements.txt
pandas>=2.2.2
numpy>=1.26.4
duckdb>=1.0.0
scikit-learn>=1.5.0
feature-engine>=1.8.0
category-encoders>=2.6.3
ydata-profiling>=4.8.3
pyyaml>=6.0.1
pydantic>=2.8.2
streamlit>=1.36.0
plotly>=5.22.0
altair>=5.3.0
